+++ tests/Type/VideoTest.php
@@ -23,6 +23,8 @@ final class VideoTest extends TestCase
         $this->assertNull($video->fileName);
         $this->assertNull($video->mimeType);
         $this->assertNull($video->fileSize);
+        $this->assertNull($video->cover);
+        $this->assertNull($video->startTimestamp);
     }
 
     public function testFromTelegramResult(): void
@@ -39,6 +41,21 @@ final class VideoTest extends TestCase
                 'width' => 500,
                 'height' => 600,
             ],
+            'cover' => [
+                [
+                    'file_id' => 'file_id-3',
+                    'file_unique_id' => 'file_unique_id',
+                    'width' => 150,
+                    'height' => 150,
+                ],
+                [
+                    'file_id' => 'file_id-4',
+                    'file_unique_id' => 'file_unique_id',
+                    'width' => 150,
+                    'height' => 150,
+                ],
+            ],
+            'start_timestamp' => 17,
             'file_name' => 'face.png',
             'mime_type' => 'image/png',
             'file_size' => 123,
@@ -53,5 +70,9 @@ final class VideoTest extends TestCase
         $this->assertSame('face.png', $video->fileName);
         $this->assertSame('image/png', $video->mimeType);
         $this->assertSame(123, $video->fileSize);
+        $this->assertCount(2, $video->cover);
+        $this->assertSame('file_id-3', $video->cover[0]->fileId);
+        $this->assertSame('file_id-4', $video->cover[1]->fileId);
+        $this->assertSame(17, $video->startTimestamp);
     }
 }